test: update to vitest
